Refactor: Moved all preferences subviews to their own class files
